requirements.txt
pika
python-telegram-bot
sqlalchemy
psycopg2